updating the dataset and judgements
The current dataset is old from 2017. all the "poster_path" in that dataset are broken. with these new changes, we can get all the correct data, with the correct poster_path, along with the movie judgments for the new dataset.

I have changed the 'wget' to 'curl', since it was giving a connection issue sometimes.